Change variable handling to import defaults first
By changing the way variable importing is handled to import the defaults
first and the custom variables second, only the settings that are
actually used/changed need to be moved to the specific settings folder.
This saves time and work and prevents the playbook from not working
after an update.