Do lots of refactoring
The playbook now starts by creating lists of enabled and disabled apps.
All tasks that are the same for every application are now done with
loops in the charjabox_deploy role. This role also imports certain
app-specific tasks.
Directories are now set in dictionaries instead of multiple variables
per app.